Added payments page with functioning button to add payments and amount. Generation of grid went to a new service.
