+++ AngularExercise/src/app/Interfaces/PaymentEntry.ts
@@ -1,0 +1,7 @@
+export interface PaymentEntry {
+    name: string,
+    amount: number;
+    code: string;
+    grid: string[][];
+    gridNumberOfCells: number;
+}
