+++ AngularExercise/src/app/app-routing.module.ts
@@ -2,6 +2,7 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { GeneratorPageComponent } from './generator-page/generator-page.component';
 import { MainPageComponent } from './main-page/main-page.component';
+import { PaymentsPageComponent } from './payments-page/payments-page.component';
 
 const routes: Routes = [
   {
@@ -11,6 +12,10 @@ const routes: Routes = [
   {
     path: 'generator',
     component: GeneratorPageComponent
+  },
+  {
+    path: 'payments',
+    component: PaymentsPageComponent
   },
   {
     path: '**',
