+++ AngularExercise/src/app/main-page/main-page.component.html
@@ -1,4 +1,4 @@
 <div class="centerDiv">
     <button [routerLink]="'/generator'">Generator</button>
-    <button>Button</button>
+    <button [routerLink]="'/payments'">Payments</button>
 </div>
