+++ AngularExercise/src/app/payments-page/payments-page.component.scss
@@ -1,0 +1,65 @@
+.justifyFlexContentCenter {
+    display: flex;
+    justify-content: center;
+}
+
+.dot {
+    height: 10px;
+    width: 10px;
+    background-color: rgba(255, 0, 0, 0.589);
+    border-radius: 50%;
+    display: inline-block;
+    vertical-align: middle;
+  }
+
+  .generatedCode {
+    margin: auto;
+    border: 2px solid darkgrey;
+    border-radius: 3px;
+    padding: 10px 30px;
+    width: fit-content;
+  }
+
+  /* Structure */
+.example-container {
+    position: relative;
+  }
+  
+  .example-table-container {
+    position: relative;
+    min-height: 100px;
+    max-height: 400px;
+    overflow: auto;
+  }
+  
+  table {
+    width: 100%;
+  }
+  
+  .example-loading-shade {
+    position: absolute;
+    top: 0;
+    left: 0;
+    bottom: 56px;
+    right: 0;
+    background: rgba(0, 0, 0, 0.15);
+    z-index: 1;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+  }
+  
+  .example-rate-limit-reached {
+    max-width: 360px;
+    text-align: center;
+  }
+  
+  /* Column Widths */
+  .mat-column-number,
+  .mat-column-state {
+    max-width: 64px;
+  }
+  
+  .mat-column-created {
+    max-width: 124px;
+  }
